Cancel Ticket & Search User

+++ app/Http/Controllers/ManagerController.php
@@ -16,12 +16,12 @@ class ManagerController extends Controller
     {
         return view('manager.home');
     }
-    public function profile()
-    {
-        $user = User::where('username', 'Mortujaii')
-            ->first();
 
-        return view('manager.profile')->with('user', $user);
+    public function profile(){
+        $user = userinfo::where('username','fahim')
+        ->first();
+        
+        return view('manager.profile')->with('user',$user);
     }
     public function editProfile(Request $req)
     {
@@ -52,6 +52,7 @@ class ManagerController extends Controller
                 'address' => $req->address
             ]);
 
+        session()->flash('msg','Profile Updated Successfully');
         return redirect()->route('manager.profile');
     }
     public function changepass(Request $req)
@@ -81,16 +82,17 @@ class ManagerController extends Controller
             $user = User::where('username', $req->uname)
                 ->update(['password' => md5($req->password)]);
 
-            $msg = "Password Changed Successfully";
+            session()->flash('msg','Password Changed Successfully');
             return redirect()->route('manager.profile');
-        } else {
-            $msg = "Wrong Old Password";
 
-            $user = User::where('username', ($req->uname))
-                ->select('id', 'password')
-                ->first();
+        }
+        else{
+            $user = userinfo::where('username',($req->uname))
+            ->select('id','password')
+            ->first();
 
-            return  redirect()->route('manager.changepass', ['id' => encrypt($user->id)]);
+            session()->flash('msg','Wrong Old Password');
+            return  redirect()->route('manager.changepass',['id'=>encrypt($user->id)]);
         }
     }
     public function userlist()
@@ -122,16 +124,22 @@ class ManagerController extends Controller
         } elseif ($req->booked == '1' && $req->fromstopage != '0' && $req->tostopage == '0') {
             $tickets = PurchasedTicket::where('from_stopage_id', '=', $req->fromstopage)->get();
             $users = array();
-            foreach ($tickets as $ticket) {
-                array_push($users, $ticket->user);
+
+            foreach($tickets as $ticket) {
+                if(!in_array($ticket->user,$users)){
+                    array_push($users,$ticket->user);
+                }
             }
             $stopage = Stopage::all();
             return view('manager.userlist')->with('users', $users)->with('stopage', $stopage);
         } elseif ($req->booked == '1' && $req->fromstopage == '0' && $req->tostopage != '0') {
             $tickets = PurchasedTicket::where('to_stopage_id', '=', $req->tostopage)->get();
             $users = array();
-            foreach ($tickets as $ticket) {
-                array_push($users, $ticket->user);
+
+            foreach($tickets as $ticket) {
+                if(!in_array($ticket->user,$users)){
+                    array_push($users,$ticket->user);
+                }
             }
             $stopage = Stopage::all();
             return view('manager.userlist')->with('users', $users)->with('stopage', $stopage);
@@ -139,8 +147,11 @@ class ManagerController extends Controller
             $tickets = PurchasedTicket::where('from_stopage_id', '=', $req->fromstopage)
                 ->where('to_stopage_id', '=', $req->tostopage)->get();
             $users = array();
-            foreach ($tickets as $ticket) {
-                array_push($users, $ticket->user);
+
+            foreach($tickets as $ticket) {
+                if(!in_array($ticket->user,$users)){
+                    array_push($users,$ticket->user);
+                }
             }
             $stopage = Stopage::all();
             return view('manager.userlist')->with('users', $users)->with('stopage', $stopage);
@@ -165,10 +176,58 @@ class ManagerController extends Controller
         //return $available_seat;
         return view('manager.flightdetails')->with('flight', $flight)->with('available_seat', $available_seat);
     }
-    public function cancelticket(Request $req)
-    {
 
-        $deleteticket = PurchasedTicket::where('id', '=', decrypt($req->id))->delete();
-        return $deleteticket;
+    public function cancelticket(Request $req){
+        
+        $ticketcount = PurchasedTicket::where('purchased_by','=',decrypt($req->uid))->count();
+        
+        if($ticketcount <= 1){
+            $user = userinfo::where('id','=',decrypt($req->uid))->first();
+
+            session()->flash('msg','Ticket cannot be canceled');
+            return view('manager.userdetails')->with('user',$user);
+        }
+        else{
+            $deleteticket = PurchasedTicket::where('id','=',decrypt($req->id))->delete();
+            $user = userinfo::where('id','=',decrypt($req->uid))->first();
+
+            session()->flash('msg','Ticket cancelled Successfully');
+            return view('manager.userdetails')->with('user',$user);
+
+        }
+
+    }
+    public function searchuserlist(){
+        $users = userinfo::where('role','User')->get();
+        return view('manager.searchuserlist')->with('users',$users);
+    }
+    public function searchuserlistsubmit(Request $req){
+
+        if($req->email == "" && $req->uname == ""){
+            $users = userinfo::where('role','User')->get();
+            return view('manager.searchuserlist')->with('users',$users);
+
+        }
+        elseif($req->email != "" && $req->uname == ""){
+            $users = userinfo::where('role','User')
+            ->where('email','like','%'.$req->email.'%')
+            ->get();
+            return view('manager.searchuserlist')->with('users',$users);
+        }
+        elseif($req->email == "" && $req->uname != ""){
+            $users = userinfo::where('role','User')
+            ->where('username','like','%'.$req->uname.'%')
+            ->get();
+            return view('manager.searchuserlist')->with('users',$users);
+        }
+        elseif($req->email != "" && $req->uname != ""){
+            $users = userinfo::where('role','User')
+            ->where('username','like','%'.$req->uname.'%')
+            ->where('email','like','%'.$req->email.'%')
+            ->get();
+            return view('manager.searchuserlist')->with('users',$users);
+        }
+
+
     }
 }
